Added functionality to acceleration and deceleration. Rudimentary detection for straight lines and curves (decelerates on curves to the right, and turn tighter on curves to the left).
Added touch led to indicate acceleration and deceleration,
and another one to indicate turning speed. These are for
debugging purposes. Port2 is for turning speed, and port5 is
for general speed. Blue means slow, green means acceleration,
red means speed is capped to the max.

Also tweaks for numbers in general.

Completes the track unreliably (see line 114, change it to
100 or 120 for more reliability). Completes it relatively fast
however (under 100 seconds).